// libs/feature-warehouses/src/models/warehouse-form-value.model.ts
export interface WarehouseFormValue {
  readonly code: string | null;
  readonly name: string | null;
  readonly description: string | null;
  readonly warehouseTypeLookupValueId: string | null;
  readonly address: string | null;
  readonly responsiblePerson: string | null;
  readonly phone: string | null;
  readonly email: string | null;
  readonly active: boolean;
}